feat: 기초 스택 코드 추가
