+++ src/type.ts
@@ -1,0 +1,5 @@
+export type RootStackParamList = {
+  Home: undefined;
+  OrderHistory: undefined;
+  Settings: undefined;
+};
